refactor: group screens by subject, and let R8 near the release build
