+++ app/src/main/java/com/aoscoremonitor/ui/screens/DescriptorsScreen.kt
@@ -1,4 +1,4 @@
-package com.aoscoremonitor.ui.screens.jni
+package com.aoscoremonitor.ui.screens
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.Arrangement
